feat(admin): migrate routing from remix-flat-routes to react-router-auto-routes
Replace remix-flat-routes + @react-router/remix-routes-option-adapter with
react-router-auto-routes. Restructure routes from flat dot-delimited files
to folder-based hierarchy with + prefix colocation for cleaner organization.
