+++ apps/admin/app/routes.ts
@@ -1,6 +1,4 @@
-import { remixRoutesOptionAdapter } from '@react-router/remix-routes-option-adapter'
-import { flatRoutes } from 'remix-flat-routes'
+import type { RouteConfig } from '@react-router/dev/routes'
+import { autoRoutes } from 'react-router-auto-routes'
 
-export default remixRoutesOptionAdapter((defineRotue) =>
-  flatRoutes('routes', defineRotue),
-)
+export default autoRoutes({ routesDir: 'routes' }) satisfies RouteConfig
