+++ apps/admin/app/routes/_admin/$city/$area/$category/$rank/+queries.server.ts
