+++ apps/admin/app/routes/_admin/$city/$area/$category/+queries.server.ts
