+++ apps/admin/app/routes/_admin/$city/$area/$category/route.tsx
