+++ apps/admin/app/routes/_admin/$city/$area/+components/category-nav-item.tsx
