+++ apps/admin/app/routes/_admin/areas/_index.tsx
@@ -15,7 +15,7 @@ import {
   TableHeader,
   TableRow,
 } from '~/components/ui'
-import type { Route } from './+types/route'
+import type { Route } from './+types/_index'
 
 export const loader = () => {
   return { areas }
