+++ apps/admin/app/routes/_admin/articles/new/+generate-article.server.ts
