+++ apps/admin/app/routes/_auth/logout.tsx
