+++ apps/admin/package.json
@@ -30,7 +30,6 @@
     "@libsql/kysely-libsql": "catalog:",
     "@paralleldrive/cuid2": "3.3.0",
     "@react-router/node": "catalog:",
-    "@react-router/remix-routes-option-adapter": "catalog:",
     "@react-router/serve": "catalog:",
     "accept-language": "catalog:",
     "ai": "catalog:",
@@ -52,7 +51,6 @@
     "react-twc": "catalog:",
     "rehype-highlight": "catalog:",
     "remark-gfm": "catalog:",
-    "remix-flat-routes": "catalog:",
     "remix-toast": "catalog:",
     "remix-utils": "catalog:",
     "sonner": "catalog:",
@@ -71,6 +69,7 @@
     "prettier": "catalog:",
     "prettier-plugin-organize-imports": "catalog:",
     "prettier-plugin-tailwindcss": "catalog:",
+    "react-router-auto-routes": "0.8.2",
     "tailwindcss": "catalog:",
     "tailwindcss-animate": "catalog:",
     "typescript": "catalog:",
